Did you finish it ? Yes . What did it cost ? Everyfuckingthing

+++ DoAnLTGT/List.h
@@ -15,7 +15,7 @@ public:
 	void HienThiList();
 	
 	// Hàm thêm
-	void Them(string, string, string, string, string, string, int, int, int);
+	void Them(string, string, string, string, string, string, int, long long int, long long int);
 	
 	// Hàm xóa
 	void XoaTheoMa(string);
@@ -29,9 +29,9 @@ public:
 	void TimKiemTheoLoai(string);
 	
 	// Hàm sắp xếp ( sử dụng con trỏ hàm )
-	void SapXepTheoSoLuong(bool(*)(int, int));
-	void SapXepTheoDonGia(bool(*)(int, int));
-	void SapXepTheoThanhTien(bool(*)(int, int));
+	void SapXepTheoSoLuong(bool(*)(long long int,long long int));
+	void SapXepTheoDonGia(bool(*)(long long int, long long int));
+	void SapXepTheoThanhTien(bool(*)(long long int, long long int));
 	
 	// Hàm ghi lại vào file
 	void GhiVaoFile();
